Refactor: relocate event auth api to its own file
